feat: Add FAQsContent and FAQAccordion components for improved FAQ handling
